+++ lib/src/core/super_keys.dart
@@ -9,4 +9,8 @@ class SuperKeys {
   static const String cartCount = 'cartCount';
   static const String cartBox = 'cartBox';
   static const String wishlistBox = 'wishlistBox';
+  static const APP_STORE_URL =
+      'https://apps.apple.com/us/app/bhavika/id1550546054';
+  static const PLAY_STORE_URL =
+      'https://play.google.com/store/apps/details?id=com.bhavikaonline.bhavika';
 }