+++ lib/src/features/auth/presentation/bloc/forgot_password_bloc/reset_password_bloc/reset_password_bloc.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
+import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
-import 'package:meta/meta.dart';
 import 'package:super_module/src/features/auth/data/models/forgot_password_pin_verification_response_model.dart';
 import 'package:super_module/src/features/auth/domain/controller/auth_login_controller.dart';