+++ lib/src/features/auth/social/presentation/social_bloc.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
+import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
-import 'package:meta/meta.dart';
 import 'package:super_module/src/features/auth/data/models/social/social_model.dart';
 import 'package:super_module/src/features/auth/social/data/repository/i_social_repository.dart';
 import 'package:super_module/super_module.dart';