+++ lib/src/features/remote_config/data/models/submodel/description_model.dart
@@ -1,0 +1,16 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'description_model.g.dart';
+
+@JsonSerializable()
+class DescriptionModel {
+  final Map<String, dynamic> forceDescriptionText;
+  final Map<String, dynamic> descriptionText;
+
+  DescriptionModel(this.forceDescriptionText, this.descriptionText);
+
+  factory DescriptionModel.fromJson(Map<String, dynamic> json) =>
+      _$DescriptionModelFromJson(json);
+
+  Map<String, dynamic> toJson() => _$DescriptionModelToJson(this);
+}