+++ lib/src/features/remote_config/firebase_version_checker.dart
@@ -1,0 +1,53 @@
+import 'dart:io';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:super_module/src/core/super_keys.dart';
+import 'package:super_module/src/core/utils/app_launchers.dart';
+
+class FirebaseVersionChecker {
+  _showVersionDialog(context) async {
+    await showDialog<String>(
+      context: context,
+      barrierDismissible: false,
+      builder: (BuildContext context) {
+        String title = "New Update Available";
+        String message =
+            "There is a newer version of app available. Would you like to update it now?";
+        String btnLabel = "Update Now";
+        String btnLabelCancel = "Later";
+        return Platform.isIOS
+            ? new CupertinoAlertDialog(
+                title: Text(title),
+                content: Text(message),
+                actions: <Widget>[
+                  TextButton(
+                    child: Text(btnLabel),
+                    onPressed: () =>
+                        AppLaunchers().launchURL(SuperKeys.APP_STORE_URL),
+                  ),
+                  TextButton(
+                    child: Text(btnLabelCancel),
+                    onPressed: () => Navigator.pop(context),
+                  ),
+                ],
+              )
+            : new AlertDialog(
+                title: Text(title),
+                content: Text(message),
+                actions: <Widget>[
+                  TextButton(
+                    child: Text(btnLabel),
+                    onPressed: () =>
+                        AppLaunchers().launchURL(SuperKeys.PLAY_STORE_URL),
+                  ),
+                  TextButton(
+                    child: Text(btnLabelCancel),
+                    onPressed: () => Navigator.pop(context),
+                  ),
+                ],
+              );
+      },
+    );
+  }
+}