+++ lib/src/features/remote_config/presentation/bloc/firebase_event.dart
@@ -1,0 +1,6 @@
+part of 'firebase_bloc.dart';
+
+@immutable
+abstract class FirebaseEvent {}
+
+class FirebaseRemoteConfigFetchEvent extends FirebaseState {}