+++ lib/super_module.dart
@@ -3,6 +3,7 @@ library super_module;
 // export 'package:super_module/src/features/auth/data/repositories/auth_remote_repository.dart';
 export 'package:super_module/src/DI/module_injection.dart';
 export 'package:super_module/src/core/extension/cap_extension.dart';
+export 'package:super_module/src/core/service/init_crash_anlytics.dart';
 export 'package:super_module/src/core/super_keys.dart';
 export 'package:super_module/src/core/utils/app_icon.dart';
 export 'package:super_module/src/core/utils/app_launchers.dart';
@@ -42,6 +43,8 @@ export 'package:super_module/src/features/cart/presentation/bloc/cart_bloc.dart'
 export 'package:super_module/src/features/chat/domain/controller/chat_controller.dart';
 export 'package:super_module/src/features/chat/domain/controller/chat_controller.dart';
 export 'package:super_module/src/features/chat/presentation/bloc/chat_bloc.dart';
+export 'package:super_module/src/features/remote_config/presentation/bloc/firebase_bloc.dart';
+export 'package:super_module/src/features/remote_config/presentation/widget/update_alert_dialog.dart';
 export 'package:super_module/src/features/user/data/model/address_model.dart';
 export 'package:super_module/src/features/user/data/model/page_limit_model.dart';
 export 'package:super_module/src/features/user/data/model/pagination_model.dart';