+++ CHANGELOG.md
@@ -79,3 +79,10 @@
 
 # [0.5.3] - 24 MAY 2022
 - Added chat
+
+# [0.5.13] - 6 JUN 2022
+- Added firebase remote config
+
+# [0.6.0] - 20 JUN 2022
+- Updated library to latest version
+- Added firebase remote config